utilize git template
This not only standardises git source retrieval, it also introduces the
correct package relations and a revision counting pkgver function.
Because of that, the version was set to be the lowest possible revision
of 0 not to indicate any update necessity.

The intermediate relics of template expansion and distribution through
the pkgrepotools are also excluded from tracking as they are generated
from tracked files.